(ios) Removing dead code + code consolidation / simplification
